mapbox

+++ db/migrate/20231127153454_create_locations.rb
@@ -3,8 +3,8 @@ class CreateLocations < ActiveRecord::Migration[7.1]
     create_table :locations do |t|
       t.string :name
       t.string :address
-      t.integer :latitude
-      t.integer :longitude
+      t.float :latitude
+      t.float :longitude
 
       t.timestamps
     end
